requirements.txt
fastapi
uvicorn
qdrant-client
rank-bm25
groq
numpy
python-dotenv